Resolved UF-15: separate annotation processors into their own module
This change introduces a new "super core" module at the centre of
the UberFire dependency graph: uberfire-annotations. The new module
exists only to contain annotation types and the other library types
those annotations depend on. It doesn't depend on any other UberFire
module.

Unfortunately, this breaks modularization a little bit, since there
is a mix of types from VFS, Workbench, and Security in this new module.